+++ core/src/main/java/org/patternfly/style/VariableScope.java
@@ -19,5 +19,7 @@ public enum VariableScope {
 
     global,
 
+    util,
+
     component
 }