+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Add
 
 - Add `TreeViewItem.disabled(boolean)`
+- Add `Variable.utilVar(String, String...)`
 
 ### Changed